requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.23.5
pandas==2.0.3
pillow==10.4.0
opencv-python==4.12.0.88
tqdm==4.67.1
scikit-learn==1.3.2
protobuf==4.25.8
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
deepface==0.0.96
torch==2.1.2
transformers==4.46.3
streamlit==1.40.1
streamlit-cropper==0.2.2